Move unittest support into testing package
